refactor（v0.2.1-optimal）：抽取 protocol 平铺测试文件
